oxfw-runtime: minor code refactoring for runtime structure
This commit refactors runtime crate so that implementation is under
lib.rs file.
