+++ services/src/bin/snd-oxfw-ctl-service.rs
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (c) 2020 Takashi Sakamoto
 use snd_firewire_ctl_services::*;
-use oxfw_runtime::runtime::OxfwRuntime;
+use oxfw_runtime::OxfwRuntime;
 
 struct OxfwServiceCmd;
